requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
rsa==4.9
service-identity==24.1.0
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
zope.interface==6.4.post2
